requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
crispy-bootstrap5==2024.2
cycler==0.12.1
defusedxml==0.7.1
Django==4.2.11
django-bootstrap-v5==1.0.11
django-crispy-forms==2.1
django-formtools==2.5.1
djangorestframework==3.14.0
ffmpeg-downloader==0.3.0
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.5.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
psutil==5.9.8
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pydub==0.25.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.1
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
supervision==0.20.0
sympy==1.12
tbb==2021.12.0
thop==0.1.1.post2209072238
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.1.45
urllib3==2.2.1
whitenoise==6.6.0
